working poi tile rendering

+++ Source/Characters/Character.h
@@ -27,7 +27,7 @@ class Character
   void setSpawnPositionY(float y_pos);
   void setSpawnPosition(float x_pos, float y_pos);
   void setVisible(bool is_visible);
-  void setDimensions(int new_width, int new_height);
+  void setDimensions(float new_width, float new_height);
   void setSpeed(int speed);
 
   void generatePathfindingMap(GameMap* game_map);
